backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
websockets==12.0
aiofiles==23.2.1
python-dotenv==1.0.0
sqlmodel==0.0.14
aiosqlite==0.19.0
jinja2==3.1.3
requests==2.31.0
pyyaml==6.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pytest==7.4.3
pytest-asyncio==0.23.2
httpx==0.26.0
email-validator==2.1.0.post1
